rename payfort_error
