+++ lib/start.rb
@@ -2,7 +2,7 @@ require "httparty"
 require "start/base_resource"
 require "start/customer"
 require "start/charge"
-require "start/errors/payfort_error"
+require "start/errors/start_error"
 require "start/errors/authentication_error"
 require "start/errors/banking_error"
 require "start/errors/request_error"
