+++ lib/start/errors/start_error.rb
